+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 # Dev
 
 - Dependency is used as a key instead of its name
+- Added default injection value
 
 ## 2.0.4